Implement author validation and error handling
- Add AutorValidator to check for duplicate authors
- Implement RegistroDuplicadoException and OperacaoNaoPermitidaException
- Update AutorController to handle exceptions and return appropriate error responses
- Enhance AutorService with validation checks and business logic
- Add new repository method to find authors by name, birth date, and nationality
- Update ErroResposta and ErrorCampo DTOs for improved error reporting

+++ src/main/java/com/adrianodeabreu/libraryapi/service/AutorService.java
@@ -1,5 +1,7 @@
 package com.adrianodeabreu.libraryapi.service;
 
+import com.adrianodeabreu.libraryapi.config.validator.AutorValidator;
+import com.adrianodeabreu.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.adrianodeabreu.libraryapi.model.Autor;
 import com.adrianodeabreu.libraryapi.repository.AutorRepository;
 import org.springframework.stereotype.Service;
@@ -13,11 +15,15 @@ public class AutorService {
 
     private final AutorRepository repository;
 
-    public AutorService(AutorRepository repository) {
+    private final AutorValidator validator;
+
+    public AutorService(AutorRepository repository, AutorValidator validator) {
         this.repository = repository;
+        this.validator = validator;
     }
 
     public void salvar(Autor autor) {
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -25,6 +31,7 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("O autor não possui um ID");
         }
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -33,6 +40,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é possível excluir um autor que possui livros cadastrados.");
+        }
         repository.delete(autor);
     }
 
@@ -50,5 +60,9 @@ public class AutorService {
        }
 
        return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return autor.getLivros().isEmpty();
     }
 }
